requirements.txt
argcomplete==1.0.0
pycurl
PyYAML
